+++ src/test/java/com/braintreegateway/unittest/ThreeDSecureLookupRequestTest.java
@@ -394,6 +394,32 @@ public class ThreeDSecureLookupRequestTest {
     }
 
     @Test
+    public void serializesWithMerchantOnRecordName() {
+        String clientData = "{\n" +
+            "  \"authorizationFingerprint\": \"auth-fingerprint\",\n" +
+            "  \"braintreeLibraryVersion\": \"braintree/web/3.44.0\",\n" +
+            "  \"dfReferenceId\": \"ABC-123\",\n" +
+            "  \"nonce\": \"FAKE-NONCE\",\n" +
+            "  \"clientMetadata\": {\n" +
+            "    \"cardinalDeviceDataCollectionTimeElapsed\": 40,\n" +
+            "    \"issuerDeviceDataCollectionResult\": true,\n" +
+            "    \"issuerDeviceDataCollectionTimeElapsed\": 413,\n" +
+            "    \"requestedThreeDSecureVersion\": \"2\",\n" +
+            "    \"sdkVersion\": \"web/3.42.0\"\n" +
+            "  }\n" +
+            "}";
+
+        ThreeDSecureLookupRequest request = new ThreeDSecureLookupRequest();
+        request.amount("10.00");
+        request.merchantOnRecordName("merchant123");
+        request.clientData(clientData);
+
+        String outputJSON = request.toJSON();
+        assertTrue(outputJSON.matches("^\\{.+\\}$"));
+        assertTrue(outputJSON.matches("^.+\"merchantOnRecordName\":\"merchant123\".+$"));
+    }
+
+    @Test
     public void serializesWithPriorAuthenticationId() {
         String clientData = "{\n" +
                 "  \"authorizationFingerprint\": \"auth-fingerprint\",\n" +